order-update
order-update completed

+++ src/controllers/Product.js
@@ -2,7 +2,11 @@ const services = require('../service/services');
 
 function createProduct(req, res, next) {
     try {
-        services.createProduct(req, res);
+        if (req.user == "admin") {
+            services.createProduct(req, res);
+        } else {
+            res.send("Not Authorised")
+        }
     } catch (err) {
         console.error(`Error while Creating`, err.message);
         next(err);
@@ -11,7 +15,11 @@ function createProduct(req, res, next) {
 
 function updateProduct(req, res, next) {
     try {
-        services.updateProduct(req, res);
+        if (req.user == "admin") {
+            services.updateProduct(req, res);
+        } else {
+            res.send("Not Authorised")
+        }
     } catch (err) {
         console.error(`Error while Login`, err.message);
         next(err);
@@ -20,7 +28,11 @@ function updateProduct(req, res, next) {
 
 function deleteProduct(req, res, next) {
     try {
-        services.deleteProduct(req, res);
+        if (req.user == "admin") {
+            services.deleteProduct(req, res);
+        } else {
+            res.send("Not Authorised")
+        }
     } catch (err) {
         console.error(`Error while Login`, err.message);
         next(err);
@@ -44,7 +56,29 @@ function getOneByName(req, res, next) {
         next(err);
     }
 }
+function orderProduct(req, res, next) {
+    try {
+        services.orderProduct(req, res);
+    } catch (err) {
+        console.error(`Error while Login`, err.message);
+        next(err);
+    }
+}
+function listOrder(req, res, next) {
+    try {
+        services.listOrder(req, res);
+    } catch (err) {
+        console.error(`Error while Login`, err.message);
+        next(err);
+    }
+}
 
 module.exports = {
-    createProduct, updateProduct, deleteProduct, getAll, getOneByName
+    createProduct,
+    updateProduct,
+    deleteProduct,
+    getAll,
+    getOneByName,
+    listOrder,
+    orderProduct
 };
